Move page-returning iterables to separate directory
