+++ src/iterables/pagewise/eth-transaction-page-iterable.ts
@@ -1,5 +1,5 @@
-import {EthTransaction} from "../eth-transaction"
-import {ITransactionSearchResponse} from "../interfaces/ethereum";
+import {EthTransaction} from "../../eth-transaction"
+import {ITransactionSearchResponse} from "../../interfaces/ethereum";
 import {IDefaultIteratorValue, ResourcePageIterable} from "./resource-page-iterable";
 
 interface IIteratorValue extends IDefaultIteratorValue<EthTransaction> {
